+++ src/main/java/com/example/backend/comment/CommentRepository.java
@@ -1,6 +1,5 @@
 package com.example.backend.comment;
 
-import com.example.backend.domain.CommentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
