--- src/main/java/com/example/backend/domain/AuthEntity.java
@@ -1,7 +1,0 @@
-//package com.example.backend.domain;
-//
-//import jakarta.persistence.Entity;
-//
-//@Entity
-//public class AuthEntity {
-//}
