+++ src/main/java/com/example/backend/playlist/PlayListRepository.java
@@ -1,6 +1,5 @@
 package com.example.backend.playlist;
 
-import com.example.backend.domain.PlaylistEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
