+++ src/main/java/com/example/backend/post/PostEntity.java
@@ -1,6 +1,6 @@
-package com.example.backend.domain;
+package com.example.backend.post;
 
-import io.swagger.v3.oas.annotations.tags.Tags;
+
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,22 +10,28 @@ import lombok.NoArgsConstructor;
 import java.time.LocalDateTime;
 
 
-@Data
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
+@Data
 @Entity
 @Builder
-@Table(name = "comment")
-public class CommentEntity {
+@Table(name = "post")
+public class PostEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long commentId;
-
     private Long postId;
+
     private Long userId;
-    private String commentContent;
+    private String postTitle;
+    private String postCategory;
+    private String thumbnailUrl;
+    private String postSummary;
+    private String postContent;
+    private String audioUrl;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
+    private Boolean isDeleted;
+
 
 }
