+++ src/main/java/com/example/backend/post/PostRepository.java
@@ -1,6 +1,5 @@
 package com.example.backend.post;
 
-import com.example.backend.domain.PostEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -8,4 +7,5 @@ import java.util.List;
 public interface PostRepository extends JpaRepository<PostEntity,Long> {
 
     List<PostEntity> findByUserId(Long userId);
+    List<PostEntity> findBypostCategory(String category);
 }
