+++ src/main/java/com/example/backend/post/PostService.java
@@ -1,8 +1,13 @@
 package com.example.backend.post;
 
 
-import com.example.backend.domain.PostEntity;
+import com.example.backend.comment.CommentEntity;
+import com.example.backend.comment.CommentRepository;
+import com.example.backend.comment.CommentRequestDto;
+import com.example.backend.comment.CommentResponseDto;
+import com.example.backend.user.UserEntity;
 import com.example.backend.user.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -20,7 +25,8 @@ import java.util.stream.Collectors;
 public class PostService {
 
     private final PostRepository postRepository;
-
+    private final CommentRepository CommentRepository;
+    private final UserRepository userRepository;
     // 게시글 작성
     @Transactional
     public Long createPost(PostDto postDto){
@@ -139,9 +145,69 @@ public class PostService {
     //게시글 삭제
     @Transactional
     public void deletePost(@PathVariable Long postId){
+        postRepository.findById(postId).orElseThrow(() ->
+                new EntityNotFoundException("Post with id " + postId + " does not exist"));
 
         postRepository.deleteById(postId);
 
+    }
+
+    //카테고리별 게시글 조회
+    @Transactional
+    public List<PostDto> getPostsByCategory(String category){
+        List<PostEntity> posts = postRepository.findBypostCategory(category);
+
+        return posts.stream()
+                .map(post -> PostDto.builder()
+                        .postId(post.getPostId())
+                        .userId(post.getUserId())
+                        .postTitle(post.getPostTitle())
+                        .postCategory(post.getPostCategory())
+                        .thumbnailUrl(post.getThumbnailUrl())
+                        .postSummary(post.getPostSummary())
+                        .postContent(post.getPostContent())
+                        .audioUrl(post.getAudioUrl())
+                        .isDeleted(post.getIsDeleted())  // 삭제 여부 추가
+                        .createdAt(post.getCreatedAt())
+                        .updatedAt(post.getUpdatedAt())  // 업데이트 시간 추가
+                        .build())
+                .collect(Collectors.toList());
+    }
+
+    // 댓글 생성
+    @Transactional
+    public CommentResponseDto createComment(
+            @PathVariable Long postId,
+            CommentRequestDto requestDto) {
+
+        //만약 id를 찾지못하면 에러를 출력
+        UserEntity user = userRepository.findById(requestDto.getUserId())
+                .orElseThrow(() -> new IllegalArgumentException("Invalid User ID"));
+        PostEntity post = postRepository.findById(postId)
+                .orElseThrow(() -> new IllegalArgumentException("Invalid Post ID"));
+
+
+        // CommentEntity 생성
+        CommentEntity comment = CommentEntity.builder()
+                .userId(user.getUserId())
+                .postId(postId)// User의 ID를 설정
+                .commentContent(requestDto.getCommentContent())
+                .createdAt(LocalDateTime.now())
+                .updatedAt(LocalDateTime.now())
+                .build();
+
+        // 댓글 저장
+        CommentRepository.save(comment);
+
+        // 저장된 댓글을 CommentResponseDto로 변환 후 반환
+        return CommentResponseDto.builder()
+                .commentId(comment.getCommentId())  // 생성된 댓글 ID
+                .postId(postId)                     // 해당 게시물 ID
+                .userId(comment.getUserId())        // 댓글 작성자 ID
+                .commentContent(comment.getCommentContent())  // 댓글 내용
+                .createdAt(comment.getCreatedAt())  // 생성 시간
+                .updatedAt(comment.getUpdatedAt())  // 수정 시간
+                .build();
     }
 
 }
