+++ src/main/java/com/example/backend/user/UserRepository.java
@@ -1,6 +1,5 @@
 package com.example.backend.user;
 
-import com.example.backend.domain.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<UserEntity,Long> {
